Refactor MemorySegmentDao
